requirements.txt
# Fathom → Markdown Scraper Dependencies
# Pin minor versions as per R-12 requirement

playwright==1.42.0
pyperclip==1.8.2
python-dotenv==1.0.1
pyyaml==6.0.1
beautifulsoup4==4.12.3
rich==13.7.0
aiohttp==3.9.3 